refactor: centralize menu data
